Removing FAB - Syncing Panels
