+++ src/containers/court/CourtPanel.js
@@ -1,17 +1,21 @@
 import React, {useEffect} from 'react';
 import {useDispatch, useSelector} from 'react-redux';
-import {requestCourts, requestPlayers} from "../../actions";
+import {displayModal, requestCourts, requestPlayers} from "../../actions";
 import CourtSelect from "../../components/court/CourtSelect";
-import {List} from 'antd';
+import {List, Layout, Button} from 'antd';
 
 import {CourtCreateModal} from '../../components/court/CourtCreateModal';
 import {CourtEditModal} from "../../components/court/CourtEditModal";
 
+const { Content, Footer } = Layout;
+
 const twoMinutesMillis = 120000;
 
 export default function CourtPanel() {
-    const { current, upcoming } = useSelector(state => state.courts);
     const dispatch = useDispatch();
+    const { current, upcoming } = useSelector(state => state.courts);
+
+    const handleAddClick = () => dispatch(displayModal());
 
     function updateScreenInformation() {
         return dispatch(requestCourts()).then(() => dispatch(requestPlayers()));
@@ -34,19 +38,28 @@ export default function CourtPanel() {
     };
 
     return (
-        <>
-            <List>
-                <div className="list-header">Current Courts</div>
+        <Layout>
+            <Content>
+                <List>
+                    <div className="list-header">Current Courts</div>
 
-                {courtsFor(current)}
+                    {courtsFor(current)}
 
-                <div className="list-header">Upcoming Courts</div>
+                    <div className="list-header">Upcoming Courts</div>
 
-                {courtsFor(upcoming)}
-            </List>
+                    {courtsFor(upcoming)}
+                </List>
+            </Content>
+
+            <Footer>
+                <Button className="w-100"
+                        onClick={handleAddClick}>
+                    Add Court
+                </Button>
+            </Footer>
 
             <CourtCreateModal/>
             <CourtEditModal/>
-        </>
+        </Layout>
     );
 }
